refactor: modularize panel base
