+++ browser/panels/panels.js
@@ -6,20 +6,7 @@ import Button from './Button'
 import Switch from './Switch'
 import FloatingUI from './FloatingUI'
 import SubmitButton from './SubmitButton'
-
-const PanelBase = props => (
-  <Entity
-    id={props.panelId}
-    color="#213033"
-    primitive="a-box"
-    width="2.3"
-    height="0.01"
-    depth="0.6"
-    rotation={{x: 60, y: props.yRotation, z: 0}}
-    position={{x: props.xDimension, y: 3.5, z: props.zDimension}}
-    key={props.panelNumber}
-    > {props.panelWidgets} </Entity>
-)
+import PanelBase from './PanelBase'
 
 const generatePanel = (xDimension, zDimension, yRotation, panelNumber, handleFunction, panelId, handleSubmit, solvedPhase1) => {
   const ModuleName = props => (
